Add console for client.webrtc.fyi.
This removes the link into the partially broken buildbot UI and
prepares for migrating the fyi master to LUCI.

I'm intentionally not adding this one to console_groups. FYI bots are
often broken and I don't want to disturb the clean green view we have
over our tree closer waterfalls. This is how the buildbot UI worked
as well: the FYI waterfall is kind of hidden.

Bug: chromium:749455
Change-Id: I6ef084be89245a7c0ea76d7cff93d4f0fde4256b
Reviewed-on: https://webrtc-review.googlesource.com/80960
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Patrik Höglund <phoglund@webrtc.org>