Regenerate .cfg files to update the generate text protos format.
As announced on luci-announce@, lucicfg has switched to a new text proto
format.

This CL regenerates the .cfg files to avoid to generate a big diff the
next time we need to update these.

It also adds "lucicfg.check_version('1.15.0')" to avoid to switch
back to older formats in case the local lucicfg is at an older version.

Bug: None
Change-Id: I27a728963bd8a3126688614400236864352f994f
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/176842
Reviewed-by: Jeremy Leconte <jleconte@webrtc.org>
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>