Allow WebRTC developers to trigger perf tests manually
It's currently not possible for regular WebRTC developers to run perf
tests on Swarming. Sometimes it's not possible to reproduce regressions
when running the perf tests on a workstation instead of Swarming. This
leads to the CL having to be rolled back and adds manual labor. Having
the ability to run perf tests on Swarming before the CL is submitted
would therefore be helpful.

Bug: None
Change-Id: I1d2f21254c1f03c618c62c7dc434ec706c831672
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/230301
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Reviewed-by: Christoffer Jansson <jansson@google.com>
Commit-Queue: Emil Lundmark <lndmrk@webrtc.org>