Declare "role/configs.validator" binding.
It defines who is allowed to call LUCI Config validation API to
validate this LUCI project's configs. This is usually done by
presubmit jobs, and thus configs.validator role is assigned to
try job task accounts.

Previously this ACL was defined in the global "config-validation"
group. It is deprecated and being replaced with per-project ACLs
defined in per-project configs (like in this CL).

There's still a global ACL to allow any googler to call
the validation API in any LUCI project they are allowed to see.
Thus the per-project binding applies only to service accounts
(they are not googlers).

Note: this CL was generated semi-automatically and reviewers are
picked automatically based on OWNERS file.

BUG=chromium:1068817
R=terelius@webrtc.org, titovartem@webrtc.org

Change-Id: I508ceaccf2abe72a4b0e031e796d28e7b78bf8ae
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/237680
Reviewed-by: Artem Titov <titovartem@webrtc.org>
Commit-Queue: Vadim Shtayura <vadimsh@google.com>